WIP removes jtockauth, extracts errorHandler to it own module, sets error to false in initalState
