+++ cypress/integration/userCanRegisterForAnAccount.feature.js
@@ -36,4 +36,22 @@ describe("User can register for an account", () => {
       cy.get("[data-cy=confirm-password-input").should("contain.text", "");
     });
   });
+
+  describe("when the registration is unsuccessful", () => {
+    describe("because of missmatched passwords", () => {
+      beforeEach(() => {
+        cy.intercept("POST", "**api/auth**", {
+          statusCode: 422,
+        });
+        cy.get("[data-cy=email-input]").type("user@email.com");
+        cy.get("[data-cy=password-input]").type("password");
+        cy.get("[data-cy=confirm-password-input").type("not_the_same_password");
+        cy.get("[data-cy=btn-signup]").click();
+      });
+
+      it.only("is expected to return a 422 status code", () => {
+
+      });
+    });
+  });
 });
