+++ package.json
@@ -8,7 +8,6 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.23.0",
     "i18next": "^21.3.2",
-    "j-tockauth": "^1.2.7",
     "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
