+++ src/modules/article.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import store from "../state/store/configureStore";
+import errorHandler from "./error"
 
 const Article = {
   async index(categoryName) {
@@ -29,19 +30,19 @@ const Article = {
   },
 };
 
-const errorHandler = (error) => {
-  if (error.response.status) {
-    store.dispatch({
-      type: "ERROR_MESSAGE",
-      payload:
-        "We are sorry! Your request can not be processed at this time. Try again later",
-    });
-  } else {
-    store.dispatch({
-      type: "ERROR_MESSAGE",
-      payload: error.message,
-    });
-  }
-};
+// const errorHandler = (error) => {
+//   if (error.response.status) {
+//     store.dispatch({
+//       type: "ERROR_MESSAGE",
+//       payload:
+//         "We are sorry! Your request can not be processed at this time. Try again later",
+//     });
+//   } else {
+//     store.dispatch({
+//       type: "ERROR_MESSAGE",
+//       payload: error.message,
+//     });
+//   }
+// };
 
 export { Article };
