+++ src/state/store/initialState.js
@@ -3,6 +3,7 @@ const initialState = {
   article: null,
   categories: [],
   authenticated: false,
+  error: false,
 };
 
 export default initialState;
